Remove Feldspar.Core.Constructs
Move Data to Reify and Syntax to Language.
